# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Shared Accounts Manager API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /stats/audit-log:
    get:
      operationId: listAuditLog
      tags: [stats]
      summary: Recent audit log entries
      parameters:
        - name: limit
          in: query
          required: false
          schema:
            type: integer
        - name: action
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Audit log entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AuditLogEntry"

  # Customers
...
  /subscriptions:
    get:
      operationId: listSubscriptions
      tags: [subscriptions]
      summary: List subscriptions with optional status or customer filter
      parameters:
        - name: status
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/SubscriptionStatus"
        - name: customerId
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: Subscriptions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SubscriptionSummary"
  /subscriptions/{id}:
    get:
      operationId: getSubscription
      tags: [subscriptions]
      summary: Get subscription detail, payments, and same-slot history
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: Subscription detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubscriptionDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /subscriptions/{id}/notes:
    patch:
      operationId: updateSubscriptionNotes
      tags: [subscriptions]
      summary: Update subscription notes
      parameters:
        - $ref: "#/components/parameters/Id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubscriptionNotesInput"
      responses:
        "200":
          description: Notes updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subscription"
  /subscriptions/{id}/cancel:
    post:
      operationId: cancelSubscription
      tags: [subscriptions]
      summary: Atomically cancel an active subscription and free its slot
      parameters:
        - $ref: "#/components/parameters/Id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CancelSubscriptionInput"
      responses:
        "200":
          description: Subscription cancelled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subscription"
        "409":
          description: Subscription is not active
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /subscriptions/{id}/refund:
    post:
      operationId: refundSubscription
      tags: [subscriptions]
      summary: Record a full refund for a cancelled subscription
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "201":
          description: Refund recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payment"
        "409":
          description: Subscription is not cancelled or has already been fully refunded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /subscriptions/{id}/renew:
    post:
      operationId: renewSubscription
      tags: [subscriptions]
      summary: Atomically renew into a new subscription with full payment
      parameters:
        - $ref: "#/components/parameters/Id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RenewalInput"
      responses:
        "201":
          description: Renewal completed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RenewalResult"
        "409":
          description: Subscription cannot be renewed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema:
        type: integer
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    AuditLogEntry:
      type: object
      required: [id, action, entity, entityId, detail, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: ["integer", "null"]
        userName:
          type: ["string", "null"]
        action:
          type: string
        entity:
          type: string
        entityId:
          type: ["integer", "null"]
        detail:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
...
    SubscriptionStatus:
      type: string
      enum: [active, expired, cancelled]
    Subscription:
      type: object
      required:
        [
          id,
          slotId,
          customerId,
          startDate,
          expiryDate,
          price,
          status,
          createdAt,
        ]
      properties:
        id:
          type: integer
        slotId:
          type: integer
        customerId:
          type: integer
        startDate:
          type: string
          format: date
        expiryDate:
          type: string
          format: date
        price:
          type: number
        status:
          $ref: "#/components/schemas/SubscriptionStatus"
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    SubscriptionSummary:
      type: object
      required:
        [
          id,
          customerId,
          customerName,
          slotId,
          slotIndex,
          accountId,
          accountLabel,
          productId,
          productName,
          startDate,
          expiryDate,
          price,
          status,
          createdAt,
        ]
      properties:
        id:
          type: integer
        customerId:
          type: integer
        customerName:
          type: string
        slotId:
          type: integer
        slotIndex:
          type: integer
        accountId:
          type: integer
        accountLabel:
          type: string
        productId:
          type: integer
        productName:
          type: string
        startDate:
          type: string
          format: date
        expiryDate:
          type: string
          format: date
        price:
          type: number
        status:
          $ref: "#/components/schemas/SubscriptionStatus"
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    Payment:
      type: object
      required: [id, subscriptionId, amount, method, paidAt]
      properties:
        id:
          type: integer
        subscriptionId:
          type: integer
        amount:
          type: number
        method:
          type: string
          enum: [cash, transfer, other]
        paidAt:
          type: string
          format: date-time
        loggedBy:
          type: ["integer", "null"]
        loggedByName:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
    SubscriptionDetail:
      allOf:
        - $ref: "#/components/schemas/SubscriptionSummary"
        - type: object
          required: [payments, slotHistory]
          properties:
            payments:
              type: array
              items:
                $ref: "#/components/schemas/Payment"
            slotHistory:
              type: array
              items:
                $ref: "#/components/schemas/SubscriptionSummary"
    SubscriptionNotesInput:
      type: object
      required: [notes]
      properties:
        notes:
          type: ["string", "null"]
    CancelSubscriptionInput:
      type: object
      required: [refunded]
      properties:
        refunded:
          type: boolean
          description: Whether the customer received a full refund when the subscription was cancelled
...
    RenewalInput:
      type: object
      required: [durationDays, price, paymentMethod, paidAt]
      properties:
        durationDays:
          type: integer
          minimum: 1
        price:
          type: number
          minimum: 0
        paymentMethod:
          type: string
          enum: [cash, transfer, other]
        paidAt:
          type: string
          format: date-time
        notes:
          type: ["string", "null"]
    RenewalResult:
      type: object
      required: [subscription, payment, previousSubscriptionId]
      properties:
        subscription:
          $ref: "#/components/schemas/Subscription"
        payment:
          $ref: "#/components/schemas/Payment"
        previousSubscriptionId:
          type: integer
...